aegis-agent/requirements.txt
# Core framework
fastapi==0.115.0
Flask==3.1.0
uvicorn[standard]==0.32.0
websockets==13.1

# LLM & AI
openai==1.100.2
httpx==0.28.1

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Data processing
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.11.7

# Document generation
Jinja2==3.1.6
python-docx==1.2.0
reportlab==4.4.3

# Config & logging
python-dotenv==1.1.1
structlog==25.4.0
PyYAML==6.0.2

# HTTP & auth
requests==2.32.5
pysmb

# Dev tools
black==25.1.0
coverage==7.10.4
flake8==7.3.0
isort==6.0.1
pylint==3.3.8
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==6.2.1
playwright==1.58.0
